+++ src/bus/mod.rs
@@ -11,7 +11,7 @@ pub use self::i2c::I2CBus;
 use crate::error::Result;
 
 pub trait DataBus {
-    fn write<D: DelayUs<u16> + DelayMs<u8>>(
+    fn write<D: DelayUs<u16> + DelayMs<u16>>(
         &mut self,
         byte: u8,
         data: bool,
